added the slate-serializes configurations in same file and updated that file's name
